+++ client/src/main/java/de/qaware/chronix/client/HelloClient.java
@@ -3,13 +3,11 @@ package de.qaware.chronix.client;
 
 import de.qaware.chronix.client.benchmark.configurator.Configurator;
 import de.qaware.chronix.client.benchmark.queryhandler.QueryHandler;
-import de.qaware.chronix.client.benchmark.queryhandler.util.CsvImporter;
+import de.qaware.chronix.client.benchmark.util.CsvImporter;
 import de.qaware.chronix.database.BenchmarkDataSource;
 import de.qaware.chronix.database.BenchmarkQuery;
 import de.qaware.chronix.database.TimeSeries;
 import de.qaware.chronix.database.TimeSeriesMetaData;
-import de.qaware.chronix.shared.QueryUtil.BenchmarkRecord;
-import de.qaware.chronix.shared.QueryUtil.ImportRecord;
 import de.qaware.chronix.shared.QueryUtil.QueryRecord;
 import de.qaware.chronix.shared.ServerConfig.TSDBInterfaceHandler;
 import de.qaware.chronix.shared.ServerConfig.ServerConfigAccessor;
