+++ client/src/main/java/de/qaware/chronix/client/benchmark/benchmarkrunner/util/TimeSeriesCounter.java
@@ -1,6 +1,6 @@
 package de.qaware.chronix.client.benchmark.benchmarkrunner.util;
 
-import de.qaware.chronix.client.benchmark.queryhandler.util.JsonTimeSeriesHandler;
+import de.qaware.chronix.client.benchmark.util.JsonTimeSeriesHandler;
 import de.qaware.chronix.database.TimeSeriesMetaData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
