+++ client/src/test/java/Client/data/ConvertCsvToJsonTest.java
@@ -1,7 +1,6 @@
 package Client.data;
 
-import de.qaware.chronix.client.benchmark.queryhandler.util.CsvImporter;
-import de.qaware.chronix.client.benchmark.queryhandler.util.JsonTimeSeriesHandler;
+import de.qaware.chronix.client.benchmark.util.CsvImporter;
 
 import java.io.File;
 import java.util.ArrayList;
