+++ client/src/test/java/Client/data/ReadCsvFilesTest.java
@@ -1,6 +1,6 @@
 package Client.data;
 
-import de.qaware.chronix.client.benchmark.queryhandler.util.CsvImporter;
+import de.qaware.chronix.client.benchmark.util.CsvImporter;
 import de.qaware.chronix.database.TimeSeries;
 import de.qaware.chronix.database.TimeSeriesPoint;
